+++ op-acceptance-tests/tests/fusaka/fusaka_test.go
@@ -1,4 +1,4 @@
-package osaka
+package fusaka
 
 import (
 	"context"
