+++ op-acceptance-tests/tests/fusaka/init_test.go
@@ -1,4 +1,4 @@
-package osaka
+package fusaka
 
 import (
 	"bytes"
